requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
fabric==3.2.2
idna==3.10
invoke==2.2.0
oauthlib==3.2.2
paramiko==3.5.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2

# Celery and Redis
celery>=5.3.0
redis>=4.5.0

# Web Scraping Dependencies
selenium>=4.11.0
webdriver-manager>=4.0.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyMuPDF>=1.23.0
